# encoding/wkt: improve unmarshalling performance