… encoding/wkt/testdata/polygon.json
… encoding/wkt/unmarshal.go